[Refactor] 챗봇 파일 수정
