+++ IRecipe/app/src/main/java/com/umcproject/irecipe/data/module/ApiModule.kt
@@ -4,10 +4,10 @@ import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import com.umcproject.irecipe.BuildConfig
 import com.umcproject.irecipe.data.remote.AppInterceptor
-import com.umcproject.irecipe.data.remote.service.aichat.AiChatDislikeService
-import com.umcproject.irecipe.data.remote.service.aichat.AiChatExpiryService
-import com.umcproject.irecipe.data.remote.service.aichat.AiChatRandomService
-import com.umcproject.irecipe.data.remote.service.aichat.AiChatRefriService
+import com.umcproject.irecipe.data.remote.service.chat.AiChatDislikeService
+import com.umcproject.irecipe.data.remote.service.chat.AiChatExpiryService
+import com.umcproject.irecipe.data.remote.service.chat.AiChatRandomService
+import com.umcproject.irecipe.data.remote.service.chat.AiChatRefriService
 import com.umcproject.irecipe.data.remote.service.login.CheckMemberService
 import com.umcproject.irecipe.data.remote.service.login.LoginService
 import com.umcproject.irecipe.data.remote.service.login.NickDuplicationService
@@ -98,7 +98,7 @@ class ApiModule {
 
     @Provides
     @Singleton
-    fun provideAiChatRefriService(retrofit: Retrofit):AiChatRefriService{
+    fun provideAiChatRefriService(retrofit: Retrofit): AiChatRefriService {
         return retrofit.create(AiChatRefriService::class.java)
     }
 
@@ -110,7 +110,7 @@ class ApiModule {
 
     @Provides
     @Singleton
-    fun provideAiChatExpiryService(retrofit: Retrofit):AiChatExpiryService{
+    fun provideAiChatExpiryService(retrofit: Retrofit): AiChatExpiryService {
         return retrofit.create(AiChatExpiryService::class.java)
     }
 
@@ -122,13 +122,13 @@ class ApiModule {
 
     @Provides
     @Singleton
-    fun provideAiChatRandomService(retrofit: Retrofit):AiChatRandomService{
+    fun provideAiChatRandomService(retrofit: Retrofit): AiChatRandomService {
         return retrofit.create(AiChatRandomService::class.java)
     }
 
     @Provides
     @Singleton
-    fun provideAiChatDislikeService(retrofit: Retrofit):AiChatDislikeService{
+    fun provideAiChatDislikeService(retrofit: Retrofit): AiChatDislikeService {
         return retrofit.create(AiChatDislikeService::class.java)
     }
 }
