+++ IRecipe/app/src/main/java/com/umcproject/irecipe/data/remote/response/chat/AiChatExpiryResponse.kt
@@ -1,8 +1,8 @@
-package com.umcproject.irecipe.data.remote.response
+package com.umcproject.irecipe.data.remote.response.chat
 
 import com.squareup.moshi.Json
 
-data class AiChatDislikeResponse(
+data class AiChatExpiryResponse(
     @field:Json(name="code")
     val code: String,
     @field:Json(name="isSuccess")
@@ -10,5 +10,10 @@ data class AiChatDislikeResponse(
     @field:Json(name="message")
     val message: String,
     @field:Json(name="result")
-    val result: String
+    val result: ExpiryResult?
+)
+
+data class ExpiryResult(
+    @field:Json(name="getResponse")
+    val gptResponse: String
 )
