+++ IRecipe/app/src/main/java/com/umcproject/irecipe/data/remote/response/chat/AiChatRefriResponse.kt
@@ -1,8 +1,8 @@
-package com.umcproject.irecipe.data.remote.response
+package com.umcproject.irecipe.data.remote.response.chat
 
 import com.squareup.moshi.Json
 
-data class AiChatExpiryResponse(
+data class AiChatRefriResponse(
     @field:Json(name="code")
     val code: String,
     @field:Json(name="isSuccess")
@@ -10,10 +10,9 @@ data class AiChatExpiryResponse(
     @field:Json(name="message")
     val message: String,
     @field:Json(name="result")
-    val result: ExpiryResult?
+    val result: RefriResult?
 )
-
-data class ExpiryResult(
-    @field:Json(name="getResponse")
+data class RefriResult(
+    @field:Json(name="gptResponse")
     val gptResponse: String
 )
