+++ IRecipe/app/src/main/java/com/umcproject/irecipe/data/remote/service/chat/AiChatDislikeService.kt
@@ -1,7 +1,7 @@
-package com.umcproject.irecipe.data.remote.service.aichat
+package com.umcproject.irecipe.data.remote.service.chat
 
-import com.umcproject.irecipe.data.remote.request.AiChatDislikeRequest
-import com.umcproject.irecipe.data.remote.response.AiChatDislikeResponse
+import com.umcproject.irecipe.data.remote.request.chat.AiChatDislikeRequest
+import com.umcproject.irecipe.data.remote.response.chat.AiChatDislikeResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
