+++ IRecipe/app/src/main/java/com/umcproject/irecipe/data/remote/service/chat/AiChatExpiryService.kt
@@ -1,6 +1,6 @@
-package com.umcproject.irecipe.data.remote.service.aichat
+package com.umcproject.irecipe.data.remote.service.chat
 
-import com.umcproject.irecipe.data.remote.response.AiChatExpiryResponse
+import com.umcproject.irecipe.data.remote.response.chat.AiChatExpiryResponse
 import retrofit2.Response
 import retrofit2.http.GET
